First fully functional Producer and Consumer
There is still a lot of work to do, but these are functional Producers and Consumers. They need to be modified to deal with timing issues because of the synchronous nature of OpenEdge. We could so some of these things in a wrapper shared library, but we'll attempt this first.